requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
cssselect==1.2.0
cssutils==2.11.1
dataframe_image==0.2.6
defusedxml==0.7.1
fastjsonschema==2.20.0
frozenlist==1.5.0
html2image==2.0.5
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lxml==5.3.0
MarkupSafe==3.0.2
mistune==3.0.2
more-itertools==10.5.0
multidict==6.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
tinycss2==1.4.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.16.0
